:art: Firebase context improve
