+++ src/context/firebaseContext.tsx
@@ -1,19 +1,9 @@
 import { createContext } from "react";
 import { initializeApp } from 'firebase/app';
-
-const firebaseConfig = {
-    apiKey: import.meta.env.VITE_API_KEY,
-    authDomain: import.meta.env.VITE_AUTH_DOMAIN,
-    databaseURL: import.meta.env.VITE_DATABASE_URL,
-    projectId: import.meta.env.VITE_PROJECT_ID,
-    storageBucket: import.meta.env.VITE_STORAGE_BUCKET,
-    messagingSenderId: import.meta.env.VITE_MESSAGING_SENDER_ID,
-    appId: import.meta.env.VITE_APP_ID,
-    measurementId: import.meta.env.VITE_MEASUREMENT_ID
-  };
+import firebaseConfig from "../firebaseConfig";
 
 const fb_app = initializeApp(firebaseConfig);
 
-const firebaseContext = createContext(fb_app);
+const FirebaseContext = createContext(fb_app);
 
-export default firebaseContext;
+export default FirebaseContext;
